Start on SkillTree classes
Note; This is just a base and many things are still subject to change or
removal, new features will be added as well.

Added the ISkillTreeObject interface, which just contains some essential
methods.

Added the SkillTreeObject (abstract) class, which contains important
information about the SkillTreeObject, as well as getters and setters
for these fields. It's constructor should not be called manually.

Added the SkillTreeBuilder class, which allows developers to create new
SkillTreeObjects without creating a whole new class and extending
SkillTreeObject. To create a new SkillTreeObject, either create a new
instance of SkillTreeBuilder or call the 'clean' method on an already
existing instance, then start setting some fields using the setters, and
finally call the 'build' method to get the SkillTreeObject.